8_Foodgram/backend/requirements.txt
Django==3.2
django-colorfield==0.9.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
flake8==3.9.2
gunicorn==20.1.0
isort==5.12.0
Pillow==9.3.0
psycopg2-binary==2.9.3
python-dotenv==1.0.0

